requirements.txt
python-dotenv
python-telegram-bot>=20.0